Moved PostsController Create action to own source.
Just shuffling the deck chairs this time.

RSpec: 530+25+11+9+9+18+5+2+5+11+5+5+9+39+24+30 (737) examples, 0
failures
RuboCop: 243 files inspected, no offenses found
