+++ spec/controllers/posts_controller/action/create_spec.rb
@@ -3,6 +3,44 @@ require 'spec_helper'
 require 'current_user_identity'
 require 'support/broadcast_success_tester'
 
+shared_examples 'a successful post that' do
+  it 'is successful' do
+    expect(subscriber).to be_successful
+    expect(subscriber).not_to be_failure
+  end
+
+  describe 'is successful, broadcasting a Newpoc::Entity::Post payload' do
+    let(:payload) { subscriber.payload_for(:success).first }
+
+    it 'which is valid' do
+      expect(payload).to be_valid
+    end
+
+    it 'with the correct field values set' do
+      acceptable_keys = [
+        :body,
+        :created_at,
+        :image_url,
+        :slug,
+        :title,
+        :updated_at
+      ]
+      expected = post_data
+                 .select { |k, _v| acceptable_keys.include? k }
+                 .merge author_name: current_user.name
+
+      expect(payload).to be_a Newpoc::Entity::Post
+      expected.each do |attrib, value|
+        expect(payload.attributes[attrib]).to eq value
+      end
+    end
+  end # describe 'is successful, broadcasting a Newpoc::Entity::Post payload'
+end # shared_examples 'a successful post that'
+
+# ############################################################################ #
+# ############################################################################ #
+# ############################################################################ #
+
 describe PostsController::Action::Create do
   let(:repo) { UserRepository.new }
   # NOTE: Old `Actions` namespace currently used here. Oops.
@@ -59,37 +97,102 @@ describe PostsController::Action::Create do
         }
       end
 
-      # it_behaves_like 'a successful post'
-      it 'is successful' do
-        expect(subscriber).to be_successful
-        expect(subscriber).not_to be_failure
+      it_behaves_like 'a successful post that'
+    end # context 'with minimal valid post data'
+
+    context 'with additional valid post data' do
+      let(:post_data) do
+        {
+          author_name: current_user.name,
+          title: 'A Title',
+          body: 'A Body',
+          image_url: 'http://example.com/image1.png'
+        }
       end
 
-      describe 'is successful, broadcasting a Newpoc::Entity::Post payload' do
-        let(:payload) { subscriber.payload_for(:success).first }
+      it_behaves_like 'a successful post that'
+    end # context 'with additional valid post data'
 
-        it 'which is valid' do
-          expect(payload).to be_valid
+    context 'with additional but invalid post data' do
+      let(:valid_data) do
+        {
+          author_name: current_user.name,
+          title: 'A Title',
+          body: 'A Body',
+          image_url: 'http://example.com/image1.png'
+        }
+      end
+      let(:invalid_data) { { bogus: 'an invalid post data attribute' } }
+      let(:post_data) { valid_data.merge invalid_data }
+
+      it_behaves_like 'a successful post that'
+
+      it 'does not include any invalid initialiser settings' do
+        payload = subscriber.payload_for(:success).first
+        invalid_data.each_key do |k|
+          expect(payload.attributes).not_to include k
+        end
+      end
+    end # context 'with additional but invalid post data'
+
+    context 'with post status set to' do
+      let(:actual) { subscriber.payload_for(:success).first }
+
+      context 'draft' do
+        let(:post_data) do
+          {
+            author_name: current_user.name,
+            title: 'A Title',
+            body: 'A Body',
+            post_status: 'draft'
+          }
         end
 
-        it 'with the correct field values set' do
-          acceptable_keys = [
-            :body,
-            :created_at,
-            :image_url,
-            :slug,
-            :title,
-            :updated_at
-          ]
-          expected = post_data.select { |k, _v| acceptable_keys.include? k }
-                     .merge author_name: current_user.name
+        it_behaves_like 'a successful post that'
 
-          expect(payload).to be_a Newpoc::Entity::Post
-          expected.each do |attrib, value|
-            expect(payload.attributes[attrib]).to eq value
-          end
+        it 'is a draft post' do
+          expect(actual).to be_draft
         end
-      end # describe 'is successful, broadcasting a ... payload'
-    end # context 'with minimal valid post data'
+      end # context 'draft'
+
+      context 'public' do
+        let(:post_data) do
+          {
+            author_name: current_user.name,
+            title: 'A Title',
+            body: 'A Body',
+            post_status: 'public'
+          }
+        end
+
+        it_behaves_like 'a successful post that'
+
+        it 'is not a draft post' do
+          expect(actual).not_to be_draft
+        end
+      end # context 'public'
+    end # context 'with post status set to'
+
+    context 'with insufficient valid post data' do
+      let(:post_data) { { body: 'A Body' } }
+
+      it 'is unsuccessful' do
+        expect(subscriber).not_to be_successful
+        expect(subscriber).to be_failure
+      end
+
+      describe 'is unsuccessful, broadcasting a payload with' do
+        let(:payload) { subscriber.payload_for(:failure).first }
+
+        it 'the expected errors' do
+          attribs = JSON.parse(payload.message).symbolize_keys
+          entity = Newpoc::Entity::Post.new attribs
+          expect(entity).not_to be_valid
+          expect(entity).to have(1).error
+          message = entity.errors.full_messages.first
+          expect(message).to eq "Title can't be blank"
+        end
+      end # describe 'is unsuccessful, broadcasting a payload with'
+    end # context 'with insufficient valid post data'
   end # context 'with a Registered User as the current user'
 end # describe PostsController::Action::Create
